fixup! fixup! Fix remaining tests
