+++ buildSrc/build.gradle.kts
@@ -5,7 +5,7 @@ repositories {
 
 dependencies {
     gradleApi()
-    implementation("org.jetbrains.kotlin.jvm", "org.jetbrains.kotlin.jvm.gradle.plugin", "1.9.0")
+    implementation("org.jetbrains.kotlin.jvm", "org.jetbrains.kotlin.jvm.gradle.plugin", "1.9.10")
     implementation("com.avast.gradle", "gradle-docker-compose-plugin", "0.17.4")
     implementation("io.gitlab.arturbosch.detekt", "detekt-gradle-plugin", "1.21.0")
 }
