+++ exposed-core/api/exposed-core.api
@@ -549,6 +549,7 @@ public final class org/jetbrains/exposed/sql/Database {
 	public final fun getVersion ()Ljava/math/BigDecimal;
 	public final fun isVersionCovers (Ljava/math/BigDecimal;)Z
 	public final fun setUseNestedTransactions (Z)V
+	public fun toString ()Ljava/lang/String;
 }
 
 public final class org/jetbrains/exposed/sql/Database$Companion {
@@ -3071,6 +3072,7 @@ public final class org/jetbrains/exposed/sql/transactions/ThreadLocalTransaction
 	public fun setDefaultMinRepetitionDelay (J)V
 	public fun setDefaultReadOnly (Z)V
 	public fun setDefaultRepetitionAttempts (I)V
+	public fun toString ()Ljava/lang/String;
 }
 
 public final class org/jetbrains/exposed/sql/transactions/ThreadLocalTransactionManagerKt {
@@ -3294,7 +3296,7 @@ public final class org/jetbrains/exposed/sql/vendors/DatabaseDialect$DefaultImpl
 	public static fun tableColumns (Lorg/jetbrains/exposed/sql/vendors/DatabaseDialect;[Lorg/jetbrains/exposed/sql/Table;)Ljava/util/Map;
 }
 
-public final class org/jetbrains/exposed/sql/vendors/DefaultKt {
+public final class org/jetbrains/exposed/sql/vendors/DatabaseDialectKt {
 	public static final fun getCurrentDialect ()Lorg/jetbrains/exposed/sql/vendors/DatabaseDialect;
 }
 
@@ -3519,6 +3521,7 @@ public class org/jetbrains/exposed/sql/vendors/MysqlDialect : org/jetbrains/expo
 	public final fun isFractionDateTimeSupported ()Z
 	public final fun isTimeZoneOffsetSupported ()Z
 	public fun setSchema (Lorg/jetbrains/exposed/sql/Schema;)Ljava/lang/String;
+	public fun tableExists (Lorg/jetbrains/exposed/sql/Table;)Z
 }
 
 public final class org/jetbrains/exposed/sql/vendors/MysqlDialect$Companion : org/jetbrains/exposed/sql/vendors/VendorDialect$DialectNameProvider {
@@ -3652,6 +3655,7 @@ public abstract class org/jetbrains/exposed/sql/vendors/VendorDialect : org/jetb
 	public fun existingPrimaryKeys ([Lorg/jetbrains/exposed/sql/Table;)Ljava/util/Map;
 	protected fun fillConstraintCacheForTables (Ljava/util/List;)V
 	public final fun filterCondition (Lorg/jetbrains/exposed/sql/Index;)Ljava/lang/String;
+	protected final fun getAllTableNamesCache ()Ljava/util/Map;
 	public final fun getAllTablesNames ()Ljava/util/List;
 	protected final fun getColumnConstraintsCache ()Ljava/util/Map;
 	public fun getDataTypeProvider ()Lorg/jetbrains/exposed/sql/vendors/DataTypeProvider;
