+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/ColumnType.kt
@@ -15,6 +15,7 @@ import java.nio.ByteBuffer
 import java.sql.Blob
 import java.sql.Clob
 import java.sql.ResultSet
+import java.sql.SQLException
 import java.util.*
 import kotlin.reflect.KClass
 import kotlin.reflect.full.isSubclassOf
@@ -467,7 +468,7 @@ class DecimalColumnType(
         is BigDecimal -> value
         is Double -> {
             if (value.isNaN()) {
-                error("Unexpected value of type Double: NaN of ${value::class.qualifiedName}")
+                throw SQLException("Unexpected value of type Double: NaN of ${value::class.qualifiedName}")
             } else {
                 value.toBigDecimal()
             }
