+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/vendors/FunctionProvider.kt
@@ -3,157 +3,6 @@ package org.jetbrains.exposed.sql.vendors
 import org.jetbrains.exposed.exceptions.UnsupportedByDialectException
 import org.jetbrains.exposed.exceptions.throwUnsupportedException
 import org.jetbrains.exposed.sql.*
-import org.jetbrains.exposed.sql.Function
-import org.jetbrains.exposed.sql.transactions.TransactionManager
-import java.nio.ByteBuffer
-import java.util.*
-import java.util.concurrent.ConcurrentHashMap
-
-/**
- * Provides definitions for all the supported SQL data types.
- * By default, definitions from the SQL standard are provided but if a vendor doesn't support a specific type, or it is
- * implemented differently, the corresponding function should be overridden.
- */
-abstract class DataTypeProvider {
-    // Numeric types
-
-    /** Numeric type for storing 1-byte integers. */
-    open fun byteType(): String = "TINYINT"
-
-    /** Numeric type for storing 1-byte unsigned integers.
-     *
-     * **Note:** If the database being used is not MySQL, MariaDB, or SQL Server, this will represent the 2-byte
-     * integer type.
-     */
-    open fun ubyteType(): String = "SMALLINT"
-
-    /** Numeric type for storing 2-byte integers. */
-    open fun shortType(): String = "SMALLINT"
-
-    /** Numeric type for storing 2-byte unsigned integers.
-     *
-     * **Note:** If the database being used is not MySQL or MariaDB, this will represent the 4-byte integer type.
-     */
-    open fun ushortType(): String = "INT"
-
-    /** Numeric type for storing 4-byte integers. */
-    open fun integerType(): String = "INT"
-
-    /** Numeric type for storing 4-byte unsigned integers.
-     *
-     * **Note:** If the database being used is not MySQL or MariaDB, this will represent the 8-byte integer type.
-     */
-    open fun uintegerType(): String = "BIGINT"
-
-    /** Numeric type for storing 4-byte integers, marked as auto-increment. */
-    open fun integerAutoincType(): String = "INT AUTO_INCREMENT"
-
-    /** Numeric type for storing 8-byte integers. */
-    open fun longType(): String = "BIGINT"
-
-    /** Numeric type for storing 8-byte unsigned integers. */
-    open fun ulongType(): String = "BIGINT"
-
-    /** Numeric type for storing 8-byte integers, and marked as auto-increment. */
-    open fun longAutoincType(): String = "BIGINT AUTO_INCREMENT"
-
-    /** Numeric type for storing 4-byte (single precision) floating-point numbers. */
-    open fun floatType(): String = "FLOAT"
-
-    /** Numeric type for storing 8-byte (double precision) floating-point numbers. */
-    open fun doubleType(): String = "DOUBLE PRECISION"
-
-    // Character types
-
-    /** Character type for storing strings of variable length up to a maximum. */
-    open fun varcharType(colLength: Int): String = "VARCHAR($colLength)"
-
-    /** Character type for storing strings of variable length.
-     * Some database (postgresql) use the same data type name to provide virtually _unlimited_ length. */
-    open fun textType(): String = "TEXT"
-
-    /** Character type for storing strings of _medium_ length. */
-    open fun mediumTextType(): String = "TEXT"
-
-    /** Character type for storing strings of variable and _large_ length. */
-    open fun largeTextType(): String = "TEXT"
-
-    // Binary data types
-
-    /** Binary type for storing binary strings of variable and _unlimited_ length. */
-    abstract fun binaryType(): String
-
-    /** Binary type for storing binary strings of a specific [length]. */
-    open fun binaryType(length: Int): String = if (length == Int.MAX_VALUE) "VARBINARY(MAX)" else "VARBINARY($length)"
-
-    /** Binary type for storing BLOBs. */
-    open fun blobType(): String = "BLOB"
-
-    /** Binary type for storing [UUID]. */
-    open fun uuidType(): String = "BINARY(16)"
-
-    @Suppress("MagicNumber")
-    open fun uuidToDB(value: UUID): Any =
-        ByteBuffer.allocate(16).putLong(value.mostSignificantBits).putLong(value.leastSignificantBits).array()
-
-    // Date/Time types
-
-    /** Data type for storing both date and time without a time zone. */
-    open fun dateTimeType(): String = "DATETIME"
-
-    /** Data type for storing both date and time with a time zone. */
-    open fun timestampWithTimeZoneType(): String = "TIMESTAMP WITH TIME ZONE"
-
-    /** Time type for storing time without a time zone. */
-    open fun timeType(): String = "TIME"
-
-    /** Data type for storing date without time or a time zone. */
-    open fun dateType(): String = "DATE"
-
-    // Boolean type
-
-    /** Data type for storing boolean values. */
-    open fun booleanType(): String = "BOOLEAN"
-
-    /** Returns the SQL representation of the specified [bool] value. */
-    open fun booleanToStatementString(bool: Boolean): String = bool.toString().uppercase()
-
-    /** Returns the boolean value of the specified SQL [value]. */
-    open fun booleanFromStringToBoolean(value: String): Boolean = value.toBoolean()
-
-    // JSON types
-
-    /** Data type for storing JSON in a non-binary text format. */
-    open fun jsonType(): String = "JSON"
-
-    /** Data type for storing JSON in a decomposed binary format. */
-    open fun jsonBType(): String =
-        throw UnsupportedByDialectException("This vendor does not support binary JSON data type", currentDialect)
-
-    // Misc.
-
-    /** Returns the SQL representation of the specified expression, for it to be used as a column default value. */
-    open fun processForDefaultValue(e: Expression<*>): String = when {
-        e is LiteralOp<*> && e.columnType is JsonColumnMarker -> if (currentDialect is H2Dialect) {
-            "$e".substringAfter("JSON ")
-        } else {
-            "'$e'"
-        }
-
-        e is LiteralOp<*> -> "$e"
-        e is Function<*> -> "$e"
-        currentDialect is MysqlDialect -> "$e"
-        currentDialect is SQLServerDialect -> "$e"
-        else -> "($e)"
-    }
-
-    open fun precessOrderByClause(queryBuilder: QueryBuilder, expression: Expression<*>, sortOrder: SortOrder) {
-        queryBuilder.append((expression as? ExpressionAlias<*>)?.alias ?: expression, " ", sortOrder.code)
-    }
-
-    /** Returns the hex-encoded value to be inserted into the database. */
-    abstract fun hexToDb(hexString: String): String
-}
 
 /**
  * Provides definitions for all the supported SQL functions.
@@ -289,7 +138,9 @@ abstract class FunctionProvider {
      * @param pattern Pattern the expression is checked against.
      * @param mode Match mode used to check the expression.
      */
-    open fun <T : String?> Expression<T>.match(pattern: String, mode: MatchMode? = null): Op<Boolean> = with(SqlExpressionBuilder) {
+    open fun <T : String?> Expression<T>.match(pattern: String, mode: MatchMode? = null): Op<Boolean> = with(
+        SqlExpressionBuilder
+    ) {
         this@match.like(pattern)
     }
 
@@ -832,507 +683,3 @@ abstract class FunctionProvider {
         }
     }
 }
-
-/**
- * Represents metadata information about a specific column.
- */
-data class ColumnMetadata(
-    /** Name of the column. */
-    val name: String,
-    /**
-     * Type of the column.
-     *
-     * @see java.sql.Types
-     */
-    val type: Int,
-    /** Whether the column if nullable or not. */
-    val nullable: Boolean,
-    /** Optional size of the column. */
-    val size: Int?,
-    /** Is the column auto increment */
-    val autoIncrement: Boolean,
-    /** Default value */
-    val defaultDbValue: String?,
-)
-
-/**
- * Represents metadata information about a specific table's primary key.
- */
-data class PrimaryKeyMetadata(
-    /** Name of the primary key. */
-    val name: String,
-    /** Names of the primary key's columns. */
-    val columnNames: List<String>
-)
-
-/**
- * Common interface for all database dialects.
- */
-@Suppress("TooManyFunctions")
-interface DatabaseDialect {
-    /** Name of this dialect. */
-    val name: String
-
-    /** Data type provider of this dialect. */
-    val dataTypeProvider: DataTypeProvider
-
-    /** Function provider of this dialect. */
-    val functionProvider: FunctionProvider
-
-    /** Returns `true` if the dialect supports the `IF EXISTS`/`IF NOT EXISTS` option when creating, altering or dropping objects, `false` otherwise. */
-    val supportsIfNotExists: Boolean get() = true
-
-    /** Returns `true` if the dialect supports the creation of sequences, `false` otherwise. */
-    val supportsCreateSequence: Boolean get() = true
-
-    /** Returns `true` if the dialect requires the use of a sequence to create an auto-increment column, `false` otherwise. */
-    val needsSequenceToAutoInc: Boolean get() = false
-
-    /** Returns the default reference option for the dialect. */
-    val defaultReferenceOption: ReferenceOption get() = ReferenceOption.RESTRICT
-
-    /** Returns `true` if the dialect requires the use of quotes when using symbols in object names, `false` otherwise. */
-    val needsQuotesWhenSymbolsInNames: Boolean get() = true
-
-    /** Returns `true` if the dialect supports returning multiple generated keys as a result of an insert operation, `false` otherwise. */
-    val supportsMultipleGeneratedKeys: Boolean
-
-    /** Returns`true` if the dialect supports returning generated keys obtained from a sequence. */
-    val supportsSequenceAsGeneratedKeys: Boolean get() = supportsCreateSequence
-    val supportsOnlyIdentifiersInGeneratedKeys: Boolean get() = false
-
-    /** Returns `true` if the dialect supports an upsert operation returning an affected-row value of 0, 1, or 2. */
-    val supportsTernaryAffectedRowValues: Boolean get() = false
-
-    /** Returns`true` if the dialect supports schema creation. */
-    val supportsCreateSchema: Boolean get() = true
-
-    /** Returns `true` if the dialect supports subqueries within a UNION/EXCEPT/INTERSECT statement */
-    val supportsSubqueryUnions: Boolean get() = false
-
-    val supportsDualTableConcept: Boolean get() = false
-
-    val supportsOrderByNullsFirstLast: Boolean get() = false
-
-    /** Returns `true` if the dialect supports window function definitions with GROUPS mode in frame clause */
-    val supportsWindowFrameGroupsMode: Boolean get() = false
-
-    val likePatternSpecialChars: Map<Char, Char?> get() = defaultLikePatternSpecialChars
-
-    /** Returns true if autoCommit should be enabled to create/drop database */
-    val requiresAutoCommitOnCreateDrop: Boolean get() = false
-
-    /** Returns the name of the current database. */
-    fun getDatabase(): String
-
-    /** Returns a list with the names of all the defined tables. */
-    fun allTablesNames(): List<String>
-
-    /** Checks if the specified table exists in the database. */
-    fun tableExists(table: Table): Boolean
-
-    /** Checks if the specified schema exists. */
-    fun schemaExists(schema: Schema): Boolean
-
-    fun checkTableMapping(table: Table): Boolean = true
-
-    /** Returns a map with the column metadata of all the defined columns in each of the specified [tables]. */
-    fun tableColumns(vararg tables: Table): Map<Table, List<ColumnMetadata>> = emptyMap()
-
-    /** Returns a map with the foreign key constraints of all the defined columns sets in each of the specified [tables]. */
-    fun columnConstraints(
-        vararg tables: Table
-    ): Map<Pair<Table, LinkedHashSet<Column<*>>>, List<ForeignKeyConstraint>> = emptyMap()
-
-    /** Returns a map with all the defined indices in each of the specified [tables]. */
-    fun existingIndices(vararg tables: Table): Map<Table, List<Index>> = emptyMap()
-
-    /** Returns a map with the primary key metadata in each of the specified [tables]. */
-    fun existingPrimaryKeys(vararg tables: Table): Map<Table, PrimaryKeyMetadata?> = emptyMap()
-
-    /** Returns `true` if the dialect supports `SELECT FOR UPDATE` statements, `false` otherwise. */
-    fun supportsSelectForUpdate(): Boolean
-
-    /** Returns `true` if the specified [e] is allowed as a default column value in the dialect, `false` otherwise. */
-    fun isAllowedAsColumnDefault(e: Expression<*>): Boolean = e is LiteralOp<*>
-
-    /** Returns the catalog name of the connection of the specified [transaction]. */
-    fun catalog(transaction: Transaction): String = transaction.connection.catalog
-
-    /** Clears any cached values. */
-    fun resetCaches()
-
-    /** Clears any cached values including schema names. */
-    fun resetSchemaCaches()
-
-    // Specific SQL statements
-
-    /** Returns the SQL command that creates the specified [index]. */
-    fun createIndex(index: Index): String
-
-    /** Returns the SQL command that drops the specified [indexName] from the specified [tableName]. */
-    fun dropIndex(tableName: String, indexName: String, isUnique: Boolean, isPartialOrFunctional: Boolean): String
-
-    /** Returns the SQL command that modifies the specified [column]. */
-    fun modifyColumn(column: Column<*>, columnDiff: ColumnDiff): List<String>
-
-    /** Returns the SQL command that adds a primary key specified [pkName] to an existing [table]. */
-    fun addPrimaryKey(table: Table, pkName: String?, vararg pkColumns: Column<*>): String
-
-    fun createDatabase(name: String) = "CREATE DATABASE IF NOT EXISTS ${name.inProperCase()}"
-
-    fun listDatabases(): String = "SHOW DATABASES"
-
-    fun dropDatabase(name: String) = "DROP DATABASE IF EXISTS ${name.inProperCase()}"
-
-    fun setSchema(schema: Schema): String = "SET SCHEMA ${schema.identifier}"
-
-    fun createSchema(schema: Schema): String = buildString {
-        append("CREATE SCHEMA IF NOT EXISTS ")
-        append(schema.identifier)
-        appendIfNotNull(" AUTHORIZATION ", schema.authorization)
-    }
-
-    fun dropSchema(schema: Schema, cascade: Boolean): String = buildString {
-        append("DROP SCHEMA IF EXISTS ", schema.identifier)
-
-        if (cascade) {
-            append(" CASCADE")
-        }
-    }
-
-    companion object {
-        private val defaultLikePatternSpecialChars = mapOf('%' to null, '_' to null)
-    }
-}
-
-sealed class ForUpdateOption(open val querySuffix: String) {
-
-    internal object NoForUpdateOption : ForUpdateOption("") {
-        override val querySuffix: String get() = error("querySuffix should not be called for NoForUpdateOption object")
-    }
-
-    object ForUpdate : ForUpdateOption("FOR UPDATE")
-
-    // https://dev.mysql.com/doc/refman/8.0/en/innodb-locking-reads.html for clarification
-    object MySQL {
-        object ForShare : ForUpdateOption("FOR SHARE")
-
-        object LockInShareMode : ForUpdateOption("LOCK IN SHARE MODE")
-    }
-
-    // https://mariadb.com/kb/en/select/#lock-in-share-modefor-update
-    object MariaDB {
-        object LockInShareMode : ForUpdateOption("LOCK IN SHARE MODE")
-    }
-
-    // https://www.postgresql.org/docs/current/sql-select.html
-    // https://www.postgresql.org/docs/12/explicit-locking.html#LOCKING-ROWS for clarification
-    object PostgreSQL {
-        enum class MODE(val statement: String) {
-            NO_WAIT("NOWAIT"), SKIP_LOCKED("SKIP LOCKED")
-        }
-
-        abstract class ForUpdateBase(
-            querySuffix: String,
-            private val mode: MODE? = null,
-            private vararg val ofTables: Table
-        ) : ForUpdateOption("") {
-            private val preparedQuerySuffix = buildString {
-                append(querySuffix)
-                ofTables.takeIf { it.isNotEmpty() }?.let { tables ->
-                    append(" OF ")
-                    tables.joinTo(this, separator = ",") { it.tableName }
-                }
-                mode?.let {
-                    append(" ${it.statement}")
-                }
-            }
-            final override val querySuffix: String = preparedQuerySuffix
-        }
-
-        class ForUpdate(
-            mode: MODE? = null,
-            vararg ofTables: Table
-        ) : ForUpdateBase("FOR UPDATE", mode, ofTables = ofTables)
-
-        open class ForNoKeyUpdate(
-            mode: MODE? = null,
-            vararg ofTables: Table
-        ) : ForUpdateBase("FOR NO KEY UPDATE", mode, ofTables = ofTables) {
-            companion object : ForNoKeyUpdate()
-        }
-
-        open class ForShare(
-            mode: MODE? = null,
-            vararg ofTables: Table
-        ) : ForUpdateBase("FOR SHARE", mode, ofTables = ofTables) {
-            companion object : ForShare()
-        }
-
-        open class ForKeyShare(
-            mode: MODE? = null,
-            vararg ofTables: Table
-        ) : ForUpdateBase("FOR KEY SHARE", mode, ofTables = ofTables) {
-            companion object : ForKeyShare()
-        }
-    }
-
-    // https://docs.oracle.com/cd/B19306_01/server.102/b14200/statements_10002.htm#i2066346
-    object Oracle {
-        object ForUpdateNoWait : ForUpdateOption("FOR UPDATE NOWAIT")
-
-        class ForUpdateWait(timeout: Int) : ForUpdateOption("FOR UPDATE WAIT $timeout")
-    }
-}
-
-/**
- * Base implementation of a vendor dialect
- */
-abstract class VendorDialect(
-    override val name: String,
-    override val dataTypeProvider: DataTypeProvider,
-    override val functionProvider: FunctionProvider
-) : DatabaseDialect {
-
-    protected val identifierManager
-        get() = TransactionManager.current().db.identifierManager
-
-    @Suppress("UnnecessaryAbstractClass")
-    abstract class DialectNameProvider(val dialectName: String)
-
-    /* Cached values */
-    private var _allTableNames: Map<String, List<String>>? = null
-    private var _allSchemaNames: List<String>? = null
-
-    /** Returns a list with the names of all the defined tables within default scheme. */
-    val allTablesNames: List<String>
-        get() {
-            val connection = TransactionManager.current().connection
-            return getAllTableNamesCache().getValue(connection.metadata { currentScheme })
-        }
-
-    private fun getAllTableNamesCache(): Map<String, List<String>> {
-        val connection = TransactionManager.current().connection
-        if (_allTableNames == null) {
-            _allTableNames = connection.metadata { tableNames }
-        }
-        return _allTableNames!!
-    }
-
-    private fun getAllSchemaNamesCache(): List<String> {
-        val connection = TransactionManager.current().connection
-        if (_allSchemaNames == null) {
-            _allSchemaNames = connection.metadata { schemaNames }
-        }
-        return _allSchemaNames!!
-    }
-
-    override val supportsMultipleGeneratedKeys: Boolean = true
-
-    override fun getDatabase(): String = catalog(TransactionManager.current())
-
-    /**
-     * Returns a list with the names of all the defined tables with schema prefixes if database supports it.
-     * This method always re-read data from DB.
-     * Using `allTablesNames` field is the preferred way.
-     */
-    override fun allTablesNames(): List<String> = TransactionManager.current().connection.metadata {
-        tableNames.getValue(currentScheme)
-    }
-
-    override fun tableExists(table: Table): Boolean {
-        val tableScheme = table.schemaName
-        val scheme = tableScheme?.inProperCase() ?: TransactionManager.current().connection.metadata { currentScheme }
-        val allTables = getAllTableNamesCache().getValue(scheme)
-        return allTables.any {
-            when {
-                tableScheme != null -> it == table.nameInDatabaseCase()
-                scheme.isEmpty() -> it == table.nameInDatabaseCaseUnquoted()
-                else -> {
-                    val sanitizedTableName = if (currentDialect is MysqlDialect || currentDialect is SQLServerDialect) {
-                        table.tableNameWithoutScheme
-                    } else {
-                        table.tableNameWithoutSchemeSanitized
-                    }
-                    val nameInDb = "$scheme.$sanitizedTableName".inProperCase()
-                    it == nameInDb
-                }
-            }
-        }
-    }
-
-    override fun schemaExists(schema: Schema): Boolean {
-        val allSchemas = getAllSchemaNamesCache()
-        return allSchemas.any { it == schema.identifier.inProperCase() }
-    }
-
-    override fun tableColumns(vararg tables: Table): Map<Table, List<ColumnMetadata>> =
-        TransactionManager.current().connection.metadata { columns(*tables) }
-
-    override fun columnConstraints(
-        vararg tables: Table
-    ): Map<Pair<Table, LinkedHashSet<Column<*>>>, List<ForeignKeyConstraint>> {
-        val constraints = HashMap<Pair<Table, LinkedHashSet<Column<*>>>, MutableList<ForeignKeyConstraint>>()
-
-        val tablesToLoad = tables.filter { !columnConstraintsCache.containsKey(it.nameInDatabaseCaseUnquoted()) }
-
-        fillConstraintCacheForTables(tablesToLoad)
-        tables.forEach { table ->
-            columnConstraintsCache[table.nameInDatabaseCaseUnquoted()].orEmpty().forEach {
-                constraints.getOrPut(table to it.from) { arrayListOf() }.add(it)
-            }
-        }
-        return constraints
-    }
-
-    override fun existingIndices(vararg tables: Table): Map<Table, List<Index>> =
-        TransactionManager.current().db.metadata { existingIndices(*tables) }
-
-    override fun existingPrimaryKeys(vararg tables: Table): Map<Table, PrimaryKeyMetadata?> =
-        TransactionManager.current().db.metadata { existingPrimaryKeys(*tables) }
-
-    private val supportsSelectForUpdate: Boolean by lazy {
-        TransactionManager.current().db.metadata { supportsSelectForUpdate }
-    }
-
-    override fun supportsSelectForUpdate(): Boolean = supportsSelectForUpdate
-
-    protected fun String.quoteIdentifierWhenWrongCaseOrNecessary(tr: Transaction): String =
-        tr.db.identifierManager.quoteIdentifierWhenWrongCaseOrNecessary(this)
-
-    protected val columnConstraintsCache: MutableMap<String, Collection<ForeignKeyConstraint>> = ConcurrentHashMap()
-
-    protected open fun fillConstraintCacheForTables(tables: List<Table>): Unit =
-        columnConstraintsCache.putAll(TransactionManager.current().db.metadata { tableConstraints(tables) })
-
-    override fun resetCaches() {
-        _allTableNames = null
-        columnConstraintsCache.clear()
-        TransactionManager.current().db.metadata { cleanCache() }
-    }
-
-    override fun resetSchemaCaches() {
-        _allSchemaNames = null
-        resetCaches()
-    }
-
-    fun filterCondition(index: Index): String? {
-        return index.filterCondition?.let {
-            when (currentDialect) {
-                is PostgreSQLDialect, is SQLServerDialect, is SQLiteDialect -> {
-                    QueryBuilder(false)
-                        .append(" WHERE ").append(it)
-                        .toString()
-                }
-
-                else -> {
-                    exposedLogger.warn("Index creation with a filter condition is not supported in ${currentDialect.name}")
-                    return null
-                }
-            }
-        } ?: ""
-    }
-
-    private fun indexFunctionToString(function: Function<*>): String {
-        val baseString = function.toString()
-        return when (currentDialect) {
-            // SQLite & Oracle do not support "." operator (with table prefix) in index expressions
-            is SQLiteDialect, is OracleDialect -> baseString.replace(Regex("""^*[^( ]*\."""), "")
-            is MysqlDialect -> if (baseString.first() != '(') "($baseString)" else baseString
-            else -> baseString
-        }
-    }
-
-    /**
-     * Uniqueness might be required for foreign key constraints.
-     *
-     * In PostgreSQL (https://www.postgresql.org/docs/current/indexes-unique.html), UNIQUE means B-tree only.
-     * Unique constraints can not be partial
-     * Unique indexes can be partial
-     */
-    override fun createIndex(index: Index): String {
-        val t = TransactionManager.current()
-        val quotedTableName = t.identity(index.table)
-        val quotedIndexName = t.db.identifierManager.cutIfNecessaryAndQuote(index.indexName)
-        val keyFields = index.columns.plus(index.functions ?: emptyList())
-        val fieldsList = keyFields.joinToString(prefix = "(", postfix = ")") {
-            when (it) {
-                is Column<*> -> t.identity(it)
-                is Function<*> -> indexFunctionToString(it)
-                // returned by existingIndices() mapping String metadata to stringLiteral()
-                is LiteralOp<*> -> it.value.toString().trim('"')
-                else -> {
-                    exposedLogger.warn("Unexpected defining key field will be passed as String: $it")
-                    it.toString()
-                }
-            }
-        }
-        val includesOnlyColumns = index.functions?.isEmpty() != false
-        val maybeFilterCondition = filterCondition(index) ?: return ""
-
-        return when {
-            // unique and no filter -> constraint, the type is not supported
-            index.unique && maybeFilterCondition.isEmpty() && includesOnlyColumns -> {
-                "ALTER TABLE $quotedTableName ADD CONSTRAINT $quotedIndexName UNIQUE $fieldsList"
-            }
-            // unique and filter -> index only, the type is not supported
-            index.unique -> {
-                "CREATE UNIQUE INDEX $quotedIndexName ON $quotedTableName $fieldsList$maybeFilterCondition"
-            }
-            // type -> can't be unique or constraint
-            index.indexType != null -> {
-                createIndexWithType(
-                    name = quotedIndexName, table = quotedTableName,
-                    columns = fieldsList, type = index.indexType, filterCondition = maybeFilterCondition
-                )
-            }
-
-            else -> {
-                "CREATE INDEX $quotedIndexName ON $quotedTableName $fieldsList$maybeFilterCondition"
-            }
-        }
-    }
-
-    protected open fun createIndexWithType(name: String, table: String, columns: String, type: String, filterCondition: String): String {
-        return "CREATE INDEX $name ON $table $columns USING $type$filterCondition"
-    }
-
-    override fun dropIndex(tableName: String, indexName: String, isUnique: Boolean, isPartialOrFunctional: Boolean): String {
-        return "ALTER TABLE ${identifierManager.quoteIfNecessary(tableName)} DROP CONSTRAINT ${identifierManager.quoteIfNecessary(indexName)}"
-    }
-
-    override fun modifyColumn(column: Column<*>, columnDiff: ColumnDiff): List<String> =
-        listOf("ALTER TABLE ${TransactionManager.current().identity(column.table)} MODIFY COLUMN ${column.descriptionDdl(true)}")
-
-    override fun addPrimaryKey(table: Table, pkName: String?, vararg pkColumns: Column<*>): String {
-        val transaction = TransactionManager.current()
-        val columns = pkColumns.joinToString(prefix = "(", postfix = ")") { transaction.identity(it) }
-        val constraint = pkName?.let { " CONSTRAINT ${identifierManager.quoteIfNecessary(it)} " } ?: " "
-        return "ALTER TABLE ${transaction.identity(table)} ADD${constraint}PRIMARY KEY $columns"
-    }
-}
-
-private val explicitDialect = ThreadLocal<DatabaseDialect?>()
-
-internal fun <T> withDialect(dialect: DatabaseDialect, body: () -> T): T {
-    return try {
-        explicitDialect.set(dialect)
-        body()
-    } finally {
-        explicitDialect.set(null)
-    }
-}
-
-/** Returns the dialect used in the current transaction, may throw an exception if there is no current transaction. */
-val currentDialect: DatabaseDialect get() = explicitDialect.get() ?: TransactionManager.current().db.dialect
-
-internal val currentDialectIfAvailable: DatabaseDialect?
-    get() = if (TransactionManager.isInitialized() && TransactionManager.currentOrNull() != null) {
-        currentDialect
-    } else {
-        null
-    }
-
-internal fun String.inProperCase(): String =
-    TransactionManager.currentOrNull()?.db?.identifierManager?.inProperCase(this@inProperCase) ?: this
