+++ gradle.properties
@@ -1,5 +1,7 @@
 org.gradle.parallel=false
 org.gradle.jvmargs=-Dfile.encoding=UTF-8
-#
+org.gradle.configuration.cache=true
+org.gradle.caching=true
+
 group=org.jetbrains.exposed
 version=0.43.0
