+++ samples/exposed-ktor/build.gradle.kts
@@ -5,9 +5,9 @@ val exposedVersion: String by project
 val h2Version: String by project
 
 plugins {
-    kotlin("jvm") version "1.8.21"
-    id("io.ktor.plugin") version "2.3.1"
-    id("org.jetbrains.kotlin.plugin.serialization") version "1.9.0"
+    kotlin("jvm") version "1.9.10"
+    id("org.jetbrains.kotlin.plugin.serialization") version "1.9.10"
+    id("io.ktor.plugin") version "2.3.4"
 }
 
 group = "org.jetbrains.exposed.samples.ktor"
