+++ samples/exposed-ktor/gradle.properties
@@ -1,5 +1,5 @@
-ktorVersion=2.3.1
-kotlinVersion=1.8.21
+ktorVersion=2.3.4
+kotlinVersion=1.8.10
 logbackVersion=1.2.11
 kotlin.code.style=official
 exposedVersion=0.43.0
